[Issue 6] Trying to work out the use of JavaMail configuration on both Java 5 and 6.  Pain.
http://code.google.com/p/hackystat-sensorbase-uh/issues/detail?id=6

+++ src/org/hackystat/sensorbase/resource/users/UserKeyGenerator.java
@@ -42,7 +42,7 @@ class UserKeyGenerator {
      while (!isUniqueKey) {
        userKey = new StringBuffer(UserKeyGenerator.USERKEY_LENGTH);
        for (int i = 0; i < UserKeyGenerator.USERKEY_LENGTH; i++) {
-         userKey.append(charset[(int) generator.nextInt(charset.length)]);
+         userKey.append(charset[generator.nextInt(charset.length)]);
        }
        isUniqueKey = !manager.isUserKey(userKey.toString());
      }
